provider: Handle VPN configuration and association states
Set the association state when VPN is waiting for user input as an
initial state after connecting the provider. Set the configuration
state (as it is declaced to be the string to connect state in VPN)
accordingly as well. Start VPN connect timeout (starts in configuration
state) with restart option to ensure that the timeout begins from the
last known configuration (connect) state.