service: Explicit VPN connect timeout, ignore in VPN agent wait
Ignore the connect timeout autostarting when connecting a VPN service
because initially the VPN is in association state in which the VPN is
waiting for the VPN agent. Separate the starting of connect timeout into
its own function __connman_service_start_connect_timeout() so provider.c
can call it when it enters configuration state.

When a VPN is waiting for user input it should not be affected by
connect timeout as the connection is not yet attempted. This may happen
if VPN resumes to association state when requiring the VPN agent for
other, e.g., encrypted private key input after credential input.

Thus, this change makes it possible to respect both timeout values that
are configurable for vpnd and connmand separately. If the timeout for
vpnd is longer than the timeout of connmand the VPN agent dialog can be
prematurely terminated as connection timeout terminates the VPN
connection. It is more clear this way that let vpnd wait the time it is
configured to wait for the VPN agent to ask for credentials, and,
therefore, separetes the responsibilities of the two daemons.