vpn: Add VPN agent use flag for plugins
Add a flag that can be used by the VPN plugins to tell the vpn_driver
whether it uses VPN agent or not. Default to using VPN agent if the flag
is omitted.

This is done to accommodate the state transition in vpn-provider when
the VPN does not utilize VPN agent.