gresolv: Add GError for hostname lookup
Add a GError for the hostname lookup call to indicate the error as the
function should return either 0 in case of error or the source id. This
fixes the wrong returns when error happens and makes it possible to
check the error. The error message is propagated from add_query() and a
generic G_RESOLV_ERROR_ADD_QUERY error is reported as the GError code.