wireguard: Fix shutdown, ensure one exit and set no agent is used
Split disconnect into wg_disconnect() and disconnect(). Use disconnect()
to properly disconnect when error occurs with the appropriate error
code. Do the same in wg_disconnect() by using the error code to be from
device removal.

The shutdown process requires, that in order to do the proper transition
to IDLE, vpn_died() is to be called. This is normally called by the task
that has been killed so simulate this behavior in WireGuard by adding a
delayed call for that. It is to be executed only once as all the cleanup
is to be done at that last step. Use a delay of 50ms to simulate exit.

Add use of the VPN_FLAG_NO_AGENT to declare that WireGuard does not use
a VPN agent. This ensures that the state transition is done according
to state machine by __vpn_provider_connect().

Add separate create/free for struct wireguard_info for VPN consistency.