util: Add conf.d support for connmand and vpnd
Add support for reading and processing of additional configuration
files. Util is best suited for both daemons to avoid adding more
internal dependencies to vpnd.

The files are saved into the list with the full path if it is given as a
parameter. Callback can be omitted for only getting the list of
configuration files with the desired suffix in the given directory.
Callback gets the full path of the config file for convenience.

Add a Glib file error converter to utils. Map the errors in G_FILE_ERROR
domain to matching errnos. Return -EBADMSG if error is NULL and -ENOTSUP
when domain is not G_FILE_ERROR.