setting: Move settings from main to setting with hash table support
Move all setting from main.c to setting.c. Add internal and external
getters for new content. Put the defines into setting.h for other
components to access. Modify main.c to support new way of using
settings, initialize settings early so the startup options can be set as
well.

The additional configs may have a small amount of keys, it is more
optimal to read only what are set. Furthermore, this avoids from
resetting values to default which have a default value, like a string
list for blacklisted interfaces, when additional config is read by going
through all the values.

Use a new configuration options struct with union field corresponding to
the option type in order to cleanup the code and make adding new config
options more straightforward. Allow to speficy callbacks for special
parsing and value checks.

The different types that are supported in the current_val and
default_val union variables:
 - bool           CONF_TYPE_BOOL
 - double/real    CONF_TYPE_DOUBLE
 - uint           CONF_TYPE_UINT
 - uint* (array)  CONF_TYPE_UINTARR
 - char*          CONF_TYPE_STR
 - char**         CONF_TYPE_STRARR
 - GHashTable     CONF_TYPE_HASHTABLE

The char* and double/real support reading as the set value type and
converting it to an uint type. This is to accommodate some current
configuration needs.

Default values set are either copied directly (bool, double/real, uint)
or when the char* is supposed to be parsed to an integer as in this case
the char* value is not directly used, nor changed from default, thus not
getting free'd. Any array needs to be initialized in
initialize_default_values(), and free'd in __connman_setting_cleanup().

Return option parsing in main, save wifi in settings only. The options
do not need to be in settings, only the "wifi" needs to be accessed by
other parts of the code.

Add a separate log call for settings, currently prints out the online
check status as before the changes. This is to be called after all
configs are processed.