systemd-resolved: Proxy tethering client DNS requests to resolved
Proxy the DNS requests from tethering clients to resolved using the
dns.c functionality. Add an UDP listener for the tethering interface to
serve the requests from that interface and use the shared/dns.c with the
default, hardcoded systemd-resolved server 127.0.0.53 by binding the
socket to localhost interface. This ensures that the limitations set in
the systemd-resolved (by allowing only localhost to connect to the
server) can be circumvented by making the packets to originate from
localhost and not getting dropped as martians or by the server itself.