dns: Split generic DNS functionality with cache cb support
Use the code from dnsproxy.c to make a generic shared dns.c that can be
used from both dnsproxy.c and dns-systemd-resolved.c. The full
functionality is utilized with the callbacks to cache with dnsproxy.c
and with systemd-resolved the need is to allow to add a listener for
tethering. This way dns-systemd-resolved.c can operate similarly to
dnsproxy.c, to proxy the DNS requests from the tethering clients to a
caching DNS server. Thus, retaining the same functionality from the
tethering client perspective.

The callbacks are meant to be set when caching is done by the component
using shared/dns.c. This is the case with dnsproxy.c that has internal
caching functionality. Therefore, keep the internal caching of
dnsproxy.c but remove all the rest DNS related code.

To make these changes complete add also shared to dnsproxy tools sources
and include dns in dnsproxy standalone tools.