wireguard: Treat initial connect failure as unreachable host
When changing networks WireGuard may connect too early and the network
is not setup yet, causing connection attempts to the resolved host to
fail. These should not be treated as a login error, which is a terminal
error for a VPN. Instead, treat these cases as unreachable host errors
which result in an increase in the connection error limit and keep the
autoconnect value untouched.

Also add a bit more debug into the plugin. This will help to identify
configuration errors in the future.